refactor(studyrooms): reorganize component imports and enhance Study Notes functionality
- Updated import paths for StudyRoomTabs and StudyNotes components to reflect new structure.
- Introduced new StudyNotesDropdown, StudyNotesList, and SearchFilterBar components for improved functionality and user experience.
- Added dialog components for renaming, deleting, and moving study notes, enhancing interaction capabilities.
- Streamlined state management and pagination for better performance and usability.
